refactor(selector4nix): move AppError conversions
