--- src/application/nar_file/error_impl.rs
@@ -1,1 +1,0 @@
-
