+++ src/application/substituter/mod.rs
@@ -1,4 +1,2 @@
 pub mod actor;
 pub mod usecase;
-
-mod error_impl;
