+++ src/domain/nar_info/model/nar_file_name.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use snafu::{Snafu, ensure};
 
-use crate::domain::common::url::Url;
+use crate::{AppError, AppErrorKind, domain::common::url::Url};
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub struct NarFileName(String);
@@ -32,6 +32,12 @@ pub enum TryNewNarFileNameError {
     ContainsSlash,
     #[snafu(display("nar file name should end with `\".nar\"` or `\".nar.{{compression}}\"`"))]
     MissingNarExtension,
+}
+
+impl From<TryNewNarFileNameError> for AppError {
+    fn from(error: TryNewNarFileNameError) -> Self {
+        Self::new(AppErrorKind::Rule, error)
+    }
 }
 
 #[cfg(test)]
